test: move c_sharp creation fixture to c_sharp folder
