+++ back/tests/core/dependency_graph/c_sharp/fixtures/empty_class.cs
