+++ back/tests/core/dependency_graph/c_sharp/fixtures/simple_constructor.cs
