+++ back/tests/core/dependency_graph/c_sharp/fixtures/small.csproj
